+++ lib/utils/special_titles.dart
@@ -1,0 +1,8 @@
+const _specialTitlesMap = {
+  'skyfs://local/fs-dac.hns/vup.hns/.internal/shared-with-me': 'Shared with me',
+  'skyfs://local/fs-dac.hns/home/.trash': 'Trash',
+};
+
+String? getSpecialTitle(String uri) {
+  return _specialTitlesMap[uri];
+}